MacItsG5: adjust bitrate type according to Consts80211
Fixes a compiler warning regarding signed/unsigned comparison

+++ src/artery/mac/MacItsG5.cc
@@ -251,7 +251,7 @@ void MacItsG5::scheduleNextMacEvent()
 	}
 }
 
-void assert80211pBitrate(int bitrate)
+void assert80211pBitrate(uint64_t bitrate)
 {
 	for (unsigned int i = 0; i < sizeof(BITRATES_80211P); i++) {
 		if (bitrate == BITRATES_80211P[i]) return;
